+++ flink-table/flink-table-api-scala/pom.xml
@@ -68,10 +68,7 @@ under the License.
 
 		<dependency>
 			<groupId>org.apache.flink</groupId>
-			<artifactId>flink-core</artifactId>
-			<version>${project.version}</version>
-			<type>test-jar</type>
-			<scope>test</scope>
+			<artifactId>flink-test-utils-junit</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.apache.flink</groupId>
