+++ flink-test-utils-parent/flink-test-utils-junit/src/main/java/org/apache/flink/core/testutils/FlinkMatchers.java
@@ -19,18 +19,21 @@
 package org.apache.flink.core.testutils;
 
 import org.hamcrest.Description;
+import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeDiagnosingMatcher;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.time.Duration;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.function.Function;
+import java.util.function.Predicate;
 
 /**
  * Some reusable hamcrest matchers for Flink.
@@ -77,6 +80,13 @@ public class FlinkMatchers {
 	 */
 	public static <T> FutureWillFailMatcher<T> futureWillCompleteExceptionally(Duration timeout) {
 		return futureWillCompleteExceptionally(Throwable.class, timeout);
+	}
+
+	/**
+	 * Checks for a {@link Throwable} that matches by class and message.
+	 */
+	public static Matcher<Throwable> containsCause(Throwable failureCause) {
+		return new ContainsCauseMatcher(failureCause);
 	}
 
 	// ------------------------------------------------------------------------
@@ -199,6 +209,62 @@ public class FlinkMatchers {
 		public void describeTo(Description description) {
 			description.appendText("A CompletableFuture that will have failed within " +
 				timeout.toMillis() + " milliseconds with: " + validationDescription);
+		}
+	}
+
+	private static final class ContainsCauseMatcher extends TypeSafeDiagnosingMatcher<Throwable> {
+
+		private final Throwable failureCause;
+
+		private ContainsCauseMatcher(Throwable failureCause) {
+			this.failureCause = failureCause;
+		}
+
+		@Override
+		protected boolean matchesSafely(Throwable throwable, Description description) {
+			final Optional<Throwable> optionalCause = findThrowable(
+				throwable,
+				cause ->
+					cause.getClass() == failureCause.getClass() &&
+						cause.getMessage().equals(failureCause.getMessage()));
+
+			if (!optionalCause.isPresent()) {
+				description
+					.appendText("The throwable ")
+					.appendValue(throwable)
+					.appendText(" does not contain the expected failure cause ")
+					.appendValue(failureCause);
+			}
+
+			return optionalCause.isPresent();
+		}
+
+		@Override
+		public void describeTo(Description description) {
+			description
+				.appendText("Expected failure cause is ")
+				.appendValue(failureCause);
+		}
+
+		// copied from flink-core to not mess up the dependency design too much, just for a little
+		// utility method
+		private static Optional<Throwable> findThrowable(
+				Throwable throwable,
+				Predicate<Throwable> predicate) {
+			if (throwable == null || predicate == null) {
+				return Optional.empty();
+			}
+
+			Throwable t = throwable;
+			while (t != null) {
+				if (predicate.test(t)) {
+					return Optional.of(t);
+				} else {
+					t = t.getCause();
+				}
+			}
+
+			return Optional.empty();
 		}
 	}
 }
